ConfigTypes: Separate from TaskConfig to avoid import cycles
