+++ pisek/jobs/parts/verdicts_eval.py
@@ -13,7 +13,8 @@
 
 from typing import Optional
 
-from pisek.config.task_config import TaskConfig, FailMode
+from pisek.config.task_config import TaskConfig
+from pisek.config.config_types import FailMode
 from pisek.jobs.parts.solution_result import Verdict, SUBTASK_SPEC, verdict_always
